==> 2. Unsupervised Sampling.ipynb <==
import os
os.environ["KMP_DUPLICATE_LIB_OK"]="TRUE" # jupyter env setup

# =========================================================================================
# Libraries
# =========================================================================================
import gc
import time
import math
import random
import warnings
warnings.filterwarnings("ignore")
import numpy as np
import pandas as pd
from tqdm.auto import tqdm

# Custom libraries
from utils.unsupervised_utils import read_data, get_neighbors, build_training_set
from utils.utils import read_config
from utils.metrics import get_pos_score, get_f2_score

%env TOKENIZERS_PARALLELISM=false # tokens in parallel
%env TRANSFORMERS_NO_ADVISORY_WARNINGS=true # Turn off performance warnings

config = read_config() # read configuration file
DATA_PATH = "../raw_data/"

# loop over the two modes "test" and "train"
for mode in [
    "test",
    "train"
]:
   
    print(f"FOR MODE - {mode}")
    
    # read data
    topics, content, correlations, _ = read_data(data_path=DATA_PATH, config_obj=config, read_mode=mode)
    
    # nearest neighbor algorithm
    topics, content = get_neighbors(topic_df=topics, content_df=content, config_obj=config)
    
    # Merge target and topic and calculate max positive score
    topics = topics.merge(correlations, how = 'inner', left_on = ['id'], right_on = ['topic_id'])
    pos_score = get_pos_score(topics['content_ids'], topics['predictions'])
    f2_score = get_f2_score(topics['content_ids'], topics['predictions'])
    print(f'Max Recall @{config["unsupervised_model"]["top_n"]}: {pos_score}')
    print(f'F2-Score @{config["unsupervised_model"]["top_n"]}: {f2_score}')
    
    # Set the index of content to id
    content.set_index('id', inplace = True)
    
    # Construct the training set for supervised training, used in the next step
    df = build_training_set(topic_df=topics, content_df=content)
    print(f'Our {mode} set has {len(df)} rows')
    
    df.to_parquet(f'./generated_files/unsupervised_{mode}.parquet')
    df.head()
    
    del df, topics, content, correlations
    gc.collect()
